% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainsplit.R
\name{trainsplit}
\alias{trainsplit}
\title{Splits a dataframe into training and test sets.}
\usage{
trainsplit(data, trainpct = 0.75, return = "list", seed = 111)
}
\arguments{
\item{data}{The data you want to split}

\item{trainpct}{Fraction of observations to go into training set. Must be > 0 and < 1. NOTE: This function "rounds up", so that at least 1 observation will go into the training and test sets.}

\item{return}{If "list", returns list containing the training and test set. If "train" or "test", returns only the training or test sets.}

\item{seed}{Sets the random seed. Note: sets seed only locally within the function.}
}
\value{
A dataframe containing either the test set, the training set, or a list containing both.
}
\description{
Splits a dataframe into training and test sets.
}
